webService-requirements.txt
Flask==2.2.3
Pillow==9.4.0
Werkzeug==2.2.3
flask-cors==3.0.10  
gunicorn==23.0.0
tensorflow
tqdm
huggingface_hub
pandas
transformers
tensorflow
